refactor(permission): rename next.ts to index.ts and adopt Pending base class
