+++ packages/opencode/src/server/routes/permission.ts
@@ -1,7 +1,7 @@
 import { Hono } from "hono"
 import { describeRoute, validator, resolver } from "hono-openapi"
 import z from "zod"
-import { PermissionNext } from "@/permission/next"
+import { PermissionNext } from "@/permission"
 import { errors } from "../error"
 import { lazy } from "../../util/lazy"
